+++ shell.nix
@@ -4,7 +4,7 @@
 }:
 #  https://input-output-hk.github.io/haskell.nix/tutorials/development/
 pkgs.haskellPackages.shellFor {
-  packages = ps : [ ps.template ];
+  packages = ps : [ ps.ctrlc ];
   buildInputs = [
         pkgs.ghcid
         pkgs.cabal-install