+++ src/Lib.hs
@@ -1,7 +1,0 @@
-module Lib
-  ( main
-  )
-where
-
-main :: IO ()
-main = putStrLn "hello, world ss"