+++ Changelog.md
@@ -1,6 +1,10 @@
-# Change log for template project
+# Change log for ctrlc project
+
+## Version 1.0.0 
+
+Add CtrlC module
 
 ## Version 0.0.0 
 
-import [template](https://github.com/jappeace/haskell-template-project).
+import [template](https://github.com/jappeace/template).